♻️🧹 separate chart components and create a chart utils file
