+++ .gitignore
@@ -1,6 +1,12 @@
+# Vite 
 .vite
+
+# Dependencies
 package-lock.json
+
+# Storybook
 storybook-static/
+
 # Logs
 logs
 *.log
@@ -10,8 +16,6 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
-node_modules
-dist
 dist-ssr
 *.local
 
@@ -26,14 +30,8 @@ dist-ssr
 *.sln
 *.sw?
 
-# Logs
-logs
-*.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-lerna-debug.log*
-.pnpm-debug.log*
+# Stores VSCode versions used for testing VSCode extensions
+.vscode-test
 
 # Diagnostic reports (https://nodejs.org/api/report.html)
 report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
@@ -146,9 +144,6 @@ dist
 
 # TernJS port file
 .tern-port
-
-# Stores VSCode versions used for testing VSCode extensions
-.vscode-test
 
 # yarn v2
 .yarn/cache
