+++ lib/CurveChart/index.jsx
@@ -1,19 +1,14 @@
 import PropTypes from 'prop-types'
-import { getNiceTickValues } from 'recharts-scale'
 
 import {
-  Bar,
   CartesianGrid,
   ResponsiveContainer,
   Tooltip,
   XAxis,
   YAxis,
   Label,
-  Legend,
   LineChart,
-  BarChart,
   Line,
-  ReferenceLine,
 } from 'recharts'
 
 import {
@@ -25,47 +20,21 @@ import {
   formatTooltip,
   mergeData,
   setChartLang,
-} from './dataFormat'
-import { CustomTooltip } from './CustomTooltip'
-import { CustomLegend } from './CustomLegend'
+} from '../utils/chart.utils'
 
-export function buildTicks(minYAxisValue, maxYAxisValue, tickCountValue) {
-  let builtTicks = false
-  if (
-    minYAxisValue !== 'auto' &&
-    maxYAxisValue !== 'auto'
-  ) {
-    builtTicks = getNiceTickValues(
-      [minYAxisValue, maxYAxisValue],
-      tickCountValue,
-      true
-    )
-  }
-  return builtTicks
-}
-
-function Chart({
+function CurveChart({
   data,
   period,
-  legend = false,
   compareData,
-  type,
   lang,
   Ylegend = 'kWh',
-  showTooltipKeys = true,
-  referenceLineData,
-  numberOfDecimals,
-  decimalSeparator,
-  maxYAxisValue = 'auto',
-  minYAxisValue = 'auto',
-  tickCountValue = 7,
-  displaced = false,
 }) {
-  const getChartType = (type, data, period, legend, compareData) => {
-    setChartLang(lang)
-    if (type === 'LINE') {
-      const mixedData = mergeData(data, compareData)
-      return (
+  setChartLang(lang)
+  const mixedData = mergeData(data, compareData)
+
+  return (
+    <div style={{ height: '450px' }}>
+      <ResponsiveContainer>
         <LineChart width={730} height={250} data={mixedData}>
           <CartesianGrid stroke="#ccc" strokeWidth={0.5} vertical={false} />
           <XAxis
@@ -85,13 +54,12 @@ function Chart({
             tickCount={8}
             tickLine={false}
             tickFormatter={(tickItem) => `${formatDecimal(tickItem)}`}
-            tick={{ fontSize: '1rem', transform: 'translate(0, 0)' }}
-          >
+            tick={{ fontSize: '1rem', transform: 'translate(0, 0)' }}>
             <Label value={Ylegend} angle={-90} position="insideLeft" fill="#969696" />
           </YAxis>
           <Tooltip
             formatter={(value) => formatTooltip(value, Ylegend)}
-            labelFormatter={(value) => formatTooltipLabel(period, value, 'lineChart', displaced)}
+            labelFormatter={(value) => formatTooltipLabel(period, value, 'lineChart')}
             contentStyle={{ fontWeight: 'bold' }}
           />
           <Line
@@ -113,97 +81,17 @@ function Chart({
             />
           )}
         </LineChart>
-      )
-    } else {
-      return (
-        <BarChart width={730} height={250} data={data.periods}>
-          <CartesianGrid stroke="#616161" strokeWidth={0.5} vertical={false} />
-          <XAxis dataKey="range" padding={{ left: 24, right: 24 }} hide />
-          <XAxis
-            dataKey="date"
-            tickFormatter={(tickItem) => formatXAxis(period, tickItem)}
-            tick={{ fontSize: '1rem', transform: 'translate(0, 8)' }}
-            padding={{ left: 24, right: 24 }}
-            scale="band"
-            xAxisId="values"
-          />
-          <YAxis
-            type="number"
-            domain={[minYAxisValue, maxYAxisValue]}
-            axisLine={false}
-            tickCount={tickCountValue}
-            width={75}
-            tickLine={false}
-            tick={{ fontSize: '1rem', transform: 'translate(0, 0)' }}
-            ticks={buildTicks(minYAxisValue, maxYAxisValue, tickCountValue)}
-            tickFormatter={(tickItem) =>
-              `${formatDecimal(tickItem, numberOfDecimals)}`
-            }
-          >
-            <Label value={Ylegend} angle={-90} position="insideLeft" fill="#969696" />
-          </YAxis>
-          <Tooltip
-            content={
-              <CustomTooltip Ylegend={Ylegend} showTooltipKeys={showTooltipKeys} />
-            }
-            cursor={{ fill: '#f2f2f2bb' }}
-          />
-
-          {legend && !referenceLineData && <Legend />}
-          {legend && referenceLineData &&
-            <Legend
-              content={<CustomLegend referenceLineData={referenceLineData} />}
-            />
-          }
-          {data.keys.map((element) => {
-            return (
-              <Bar
-                key={element}
-                stackId="current"
-                dataKey={element}
-                fill={data.fills[element]}
-              />
-            )
-          })}
-
-          {referenceLineData && referenceLineData.map((element, index) => {
-            return (
-              <ReferenceLine
-                key={index}
-                y={element.value}
-                stroke={element.color}
-                strokeDasharray={element.stroke}
-                strokeWidth={element.strokeWidth}
-                text={element.text}
-              />
-            )
-          })
-          }
-
-        </BarChart>
-      )
-    }
-  }
-
-  return (
-    <div style={{ height: '450px' }}>
-      <ResponsiveContainer>
-        {getChartType(type, data, period, legend, compareData)}
       </ResponsiveContainer>
     </div>
   )
 }
 
-Chart.propTypes = {
+CurveChart.propTypes = {
   data: PropTypes.oneOfType([PropTypes.array, PropTypes.object]),
   period: PropTypes.string,
-  legend: PropTypes.oneOfType([PropTypes.bool, PropTypes.any]),
   compareData: PropTypes.oneOfType([PropTypes.array]),
-  type: PropTypes.string,
   lang: PropTypes.string,
   Ylegend: PropTypes.string,
-  showTooltipKeys: PropTypes.bool,
-  referenceLineData: PropTypes.oneOfType([PropTypes.array]),
 }
 
-export default Chart
+export default CurveChart
