--- lib/CurveChart/mockData/compareData.js
@@ -1,675 +1,0 @@
-const data = [
-  {
-    date: 1642374000000,
-    value: 20,
-  },
-  {
-    date: 1642377600000,
-    value: 20,
-  },
-  {
-    date: 1642381200000,
-    value: 20,
-  },
-  {
-    date: 1642384800000,
-    value: 20,
-  },
-  {
-    date: 1642388400000,
-    value: 20,
-  },
-  {
-    date: 1642392000000,
-    value: 20,
-  },
-  {
-    date: 1642395600000,
-    value: 20,
-  },
-  {
-    date: 1642399200000,
-    value: 20,
-  },
-  {
-    date: 1642402800000,
-    value: 19,
-  },
-  {
-    date: 1642406400000,
-    value: 20,
-  },
-  {
-    date: 1642410000000,
-    value: 19,
-  },
-  {
-    date: 1642413600000,
-    value: 20,
-  },
-  {
-    date: 1642417200000,
-    value: 19,
-  },
-  {
-    date: 1642420800000,
-    value: 20,
-  },
-  {
-    date: 1642424400000,
-    value: 19,
-  },
-  {
-    date: 1642428000000,
-    value: 20,
-  },
-  {
-    date: 1642431600000,
-    value: 20,
-  },
-  {
-    date: 1642435200000,
-    value: 20,
-  },
-  {
-    date: 1642438800000,
-    value: 20,
-  },
-  {
-    date: 1642442400000,
-    value: 20,
-  },
-  {
-    date: 1642446000000,
-    value: 20,
-  },
-  {
-    date: 1642449600000,
-    value: 19,
-  },
-  {
-    date: 1642453200000,
-    value: 20,
-  },
-  {
-    date: 1642456800000,
-    value: 20,
-  },
-  {
-    date: 1642460400000,
-    value: 20,
-  },
-  {
-    date: 1642464000000,
-    value: 21,
-  },
-  {
-    date: 1642467600000,
-    value: 19,
-  },
-  {
-    date: 1642471200000,
-    value: 20,
-  },
-  {
-    date: 1642474800000,
-    value: 20,
-  },
-  {
-    date: 1642478400000,
-    value: 20,
-  },
-  {
-    date: 1642482000000,
-    value: 20,
-  },
-  {
-    date: 1642485600000,
-    value: 20,
-  },
-  {
-    date: 1642489200000,
-    value: 20,
-  },
-  {
-    date: 1642492800000,
-    value: 19,
-  },
-  {
-    date: 1642496400000,
-    value: 20,
-  },
-  {
-    date: 1642500000000,
-    value: 19,
-  },
-  {
-    date: 1642503600000,
-    value: 20,
-  },
-  {
-    date: 1642507200000,
-    value: 20,
-  },
-  {
-    date: 1642510800000,
-    value: 19,
-  },
-  {
-    date: 1642514400000,
-    value: 20,
-  },
-  {
-    date: 1642518000000,
-    value: 20,
-  },
-  {
-    date: 1642521600000,
-    value: 21,
-  },
-  {
-    date: 1642525200000,
-    value: 20,
-  },
-  {
-    date: 1642528800000,
-    value: 20,
-  },
-  {
-    date: 1642532400000,
-    value: 20,
-  },
-  {
-    date: 1642536000000,
-    value: 20,
-  },
-  {
-    date: 1642539600000,
-    value: 19,
-  },
-  {
-    date: 1642543200000,
-    value: 20,
-  },
-  {
-    date: 1642546800000,
-    value: 20,
-  },
-  {
-    date: 1642550400000,
-    value: 20,
-  },
-  {
-    date: 1642554000000,
-    value: 20,
-  },
-  {
-    date: 1642557600000,
-    value: 20,
-  },
-  {
-    date: 1642561200000,
-    value: 20,
-  },
-  {
-    date: 1642564800000,
-    value: 20,
-  },
-  {
-    date: 1642568400000,
-    value: 20,
-  },
-  {
-    date: 1642572000000,
-    value: 20,
-  },
-  {
-    date: 1642575600000,
-    value: 19,
-  },
-  {
-    date: 1642579200000,
-    value: 20,
-  },
-  {
-    date: 1642582800000,
-    value: 19,
-  },
-  {
-    date: 1642586400000,
-    value: 20,
-  },
-  {
-    date: 1642590000000,
-    value: 19,
-  },
-  {
-    date: 1642593600000,
-    value: 20,
-  },
-  {
-    date: 1642597200000,
-    value: 19,
-  },
-  {
-    date: 1642600800000,
-    value: 20,
-  },
-  {
-    date: 1642604400000,
-    value: 17,
-  },
-  {
-    date: 1642608000000,
-    value: 6,
-  },
-  {
-    date: 1642611600000,
-    value: 7,
-  },
-  {
-    date: 1642615200000,
-    value: 6,
-  },
-  {
-    date: 1642618800000,
-    value: 7,
-  },
-  {
-    date: 1642622400000,
-    value: 7,
-  },
-  {
-    date: 1642626000000,
-    value: 8,
-  },
-  {
-    date: 1642629600000,
-    value: 7,
-  },
-  {
-    date: 1642633200000,
-    value: 6,
-  },
-  {
-    date: 1642636800000,
-    value: 6,
-  },
-  {
-    date: 1642640400000,
-    value: 5,
-  },
-  {
-    date: 1642644000000,
-    value: 6,
-  },
-  {
-    date: 1642647600000,
-    value: 6,
-  },
-  {
-    date: 1642651200000,
-    value: 7,
-  },
-  {
-    date: 1642654800000,
-    value: 6,
-  },
-  {
-    date: 1642658400000,
-    value: 6,
-  },
-  {
-    date: 1642662000000,
-    value: 6,
-  },
-  {
-    date: 1642665600000,
-    value: 7,
-  },
-  {
-    date: 1642669200000,
-    value: 8,
-  },
-  {
-    date: 1642672800000,
-    value: 7,
-  },
-  {
-    date: 1642676400000,
-    value: 6,
-  },
-  {
-    date: 1642680000000,
-    value: 7,
-  },
-  {
-    date: 1642683600000,
-    value: 7,
-  },
-  {
-    date: 1642687200000,
-    value: 6,
-  },
-  {
-    date: 1642690800000,
-    value: 7,
-  },
-  {
-    date: 1642694400000,
-    value: 7,
-  },
-  {
-    date: 1642698000000,
-    value: 7,
-  },
-  {
-    date: 1642701600000,
-    value: 6,
-  },
-  {
-    date: 1642705200000,
-    value: 6,
-  },
-  {
-    date: 1642708800000,
-    value: 6,
-  },
-  {
-    date: 1642712400000,
-    value: 5,
-  },
-  {
-    date: 1642716000000,
-    value: 6,
-  },
-  {
-    date: 1642719600000,
-    value: 6,
-  },
-  {
-    date: 1642723200000,
-    value: 6,
-  },
-  {
-    date: 1642726800000,
-    value: 5,
-  },
-  {
-    date: 1642730400000,
-    value: 8,
-  },
-  {
-    date: 1642734000000,
-    value: 6,
-  },
-  {
-    date: 1642737600000,
-    value: 6,
-  },
-  {
-    date: 1642741200000,
-    value: 7,
-  },
-  {
-    date: 1642744800000,
-    value: 7,
-  },
-  {
-    date: 1642748400000,
-    value: 7,
-  },
-  {
-    date: 1642752000000,
-    value: 6,
-  },
-  {
-    date: 1642755600000,
-    value: 6,
-  },
-  {
-    date: 1642759200000,
-    value: 7,
-  },
-  {
-    date: 1642762800000,
-    value: 7,
-  },
-  {
-    date: 1642766400000,
-    value: 8,
-  },
-  {
-    date: 1642770000000,
-    value: 6,
-  },
-  {
-    date: 1642773600000,
-    value: 6,
-  },
-  {
-    date: 1642777200000,
-    value: 7,
-  },
-  {
-    date: 1642780800000,
-    value: 8,
-  },
-  {
-    date: 1642784400000,
-    value: 7,
-  },
-  {
-    date: 1642788000000,
-    value: 7,
-  },
-  {
-    date: 1642791600000,
-    value: 7,
-  },
-  {
-    date: 1642795200000,
-    value: 9,
-  },
-  {
-    date: 1642798800000,
-    value: 7,
-  },
-  {
-    date: 1642802400000,
-    value: 7,
-  },
-  {
-    date: 1642806000000,
-    value: 6,
-  },
-  {
-    date: 1642809600000,
-    value: 7,
-  },
-  {
-    date: 1642813200000,
-    value: 6,
-  },
-  {
-    date: 1642816800000,
-    value: 7,
-  },
-  {
-    date: 1642820400000,
-    value: 7,
-  },
-  {
-    date: 1642824000000,
-    value: 7,
-  },
-  {
-    date: 1642827600000,
-    value: 8,
-  },
-  {
-    date: 1642831200000,
-    value: 6,
-  },
-  {
-    date: 1642834800000,
-    value: 7,
-  },
-  {
-    date: 1642838400000,
-    value: 8,
-  },
-  {
-    date: 1642842000000,
-    value: 9,
-  },
-  {
-    date: 1642845600000,
-    value: 6,
-  },
-  {
-    date: 1642849200000,
-    value: 7,
-  },
-  {
-    date: 1642852800000,
-    value: 6,
-  },
-  {
-    date: 1642856400000,
-    value: 6,
-  },
-  {
-    date: 1642860000000,
-    value: 6,
-  },
-  {
-    date: 1642863600000,
-    value: 6,
-  },
-  {
-    date: 1642867200000,
-    value: 6,
-  },
-  {
-    date: 1642870800000,
-    value: 6,
-  },
-  {
-    date: 1642874400000,
-    value: 6,
-  },
-  {
-    date: 1642878000000,
-    value: 6,
-  },
-  {
-    date: 1642881600000,
-    value: 6,
-  },
-  {
-    date: 1642885200000,
-    value: 7,
-  },
-  {
-    date: 1642888800000,
-    value: 6,
-  },
-  {
-    date: 1642892400000,
-    value: 6,
-  },
-  {
-    date: 1642896000000,
-    value: 6,
-  },
-  {
-    date: 1642899600000,
-    value: 6,
-  },
-  {
-    date: 1642903200000,
-    value: 5,
-  },
-  {
-    date: 1642906800000,
-    value: 6,
-  },
-  {
-    date: 1642910400000,
-    value: 6,
-  },
-  {
-    date: 1642914000000,
-    value: 6,
-  },
-  {
-    date: 1642917600000,
-    value: 6,
-  },
-  {
-    date: 1642921200000,
-    value: 6,
-  },
-  {
-    date: 1642924800000,
-    value: 6,
-  },
-  {
-    date: 1642928400000,
-    value: 6,
-  },
-  {
-    date: 1642932000000,
-    value: 6,
-  },
-  {
-    date: 1642935600000,
-    value: 6,
-  },
-  {
-    date: 1642939200000,
-    value: 6,
-  },
-  {
-    date: 1642942800000,
-    value: 6,
-  },
-  {
-    date: 1642946400000,
-    value: 6,
-  },
-  {
-    date: 1642950000000,
-    value: 6,
-  },
-  {
-    date: 1642953600000,
-    value: 6,
-  },
-  {
-    date: 1642957200000,
-    value: 6,
-  },
-  {
-    date: 1642960800000,
-    value: 6,
-  },
-  {
-    date: 1642964400000,
-    value: 6,
-  },
-  {
-    date: 1642968000000,
-    value: 6,
-  },
-  {
-    date: 1642971600000,
-    value: 6,
-  },
-  {
-    date: 1642975200000,
-    value: 6,
-  },
-]
-export default data
