+++ lib/SummaryBarChart/CustomLegend.jsx
