+++ lib/SummaryBarChart/CustomTooltip.jsx
