+++ lib/index.js
@@ -3,7 +3,8 @@ import Loading from './Loading'
 import DizzyError from './DizzyError'
 import TableEditor from './TableEditor'
 import SnackbarMessages from './SnackbarMessages'
-import Chart from './CurveChart'
+import CurveChart from './CurveChart'
+import SummaryPeriodChart from './SummaryBarChart'
 import ConsumptionDisplay from './ConsumptionDisplay'
 import SomDatePicker from './SomDatePicker'
 import SummaryPricesDisplay from './SummaryPricesDisplay'
@@ -14,16 +15,14 @@ import LocalStorage from './hooks/LocalStorage'
 import messages from './services/messages'
 import { registerSomEnergiaI18n } from './i18n'
 
-// TODO: posible organization:
-//                             import { Button } from '@somenergia/somenergia-ui'
-//                             import { ConsumptionDisplay } from '@somenergia/somenergia-ui/energy'
 export {
   Firebug,
   Loading,
   TableEditor,
   SnackbarMessages,
-  Chart,
   DizzyError,
+  CurveChart,
+  SummaryPeriodChart,
   ConsumptionDisplay,
   SomDatePicker,
   SummaryPricesDisplay,
