+++ lib/utils/chart.utils.js
@@ -1,3 +1,4 @@
+import { getNiceTickValues } from 'recharts-scale'
 import dayjs from 'dayjs'
 import weekday from 'dayjs/plugin/weekday'
 import localeData from 'dayjs/plugin/localeData'
@@ -8,6 +9,12 @@ import 'dayjs/locale/ca'
 
 dayjs.extend(weekday)
 dayjs.extend(localeData)
+
+export function buildTicks(minYAxisValue, maxYAxisValue, tickCountValue) {
+  return minYAxisValue !== 'auto' && maxYAxisValue !== 'auto'
+    ? getNiceTickValues([minYAxisValue, maxYAxisValue], tickCountValue, true)
+    : false
+}
 
 export const setChartLang = (lang) => {
   switch (lang) {
@@ -27,7 +34,7 @@ export const setChartLang = (lang) => {
 }
 
 export const formatNumber = (num) => {
-  return num.toLocaleString('es-ES')
+  return num.toLocaleString('es-ES');
 }
 
 export const formatTooltip = (value, Ylegend = 'kWh') => {
@@ -55,7 +62,7 @@ export const formatTooltipLabel = (period, values, type = 'barChart', displaced 
   // If true, values are final. Ex: for hour 1 label should be '00h - 01h'
   // If false, values are initial. Ex: for hour 1 label should be '01h - 02h'
   const formatDay = (values) => dayjs(values).format('DD/MM/YYYY')
-  const hour = (values, addition=0) => dayjs(values).add(addition, 'hour').format('HH')
+  const hour = (values, addition = 0) => dayjs(values).add(addition, 'hour').format('HH')
   let initialHour, finalHour
 
   const formatedDay = formatDay(values)
@@ -105,15 +112,15 @@ export const domainFromData = (data, period) => {
 export const mergeData = (arrData1 = [], arrData2 = []) => {
   return arrData1 > arrData2
     ? arrData1.map((item, index) => ({
-        date: item.date,
-        value: item.value,
-        compValue: arrData2[index]?.value,
-      }))
+      date: item.date,
+      value: item.value,
+      compValue: arrData2[index]?.value,
+    }))
     : arrData2.map((item, index) => ({
-        date: item.date,
-        value: arrData1[index]?.value,
-        compValue: item.value,
-      }))
+      date: item.date,
+      value: arrData1[index]?.value,
+      compValue: item.value,
+    }))
 }
 
 export const ticksFromData = (data, period) => {
