+++ lib/utils/utils.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { buildTicks } from './index'
+import { buildTicks } from './chart.utils'
 
 describe('buildTicks', () => {
   it('should return false when minYAxisValue is auto', () => {
