+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@somenergia/somenergia-ui",
   "private": false,
-  "version": "1.1.0",
+  "version": "1.0.0-alpha.1",
   "description": "Common React components for Som Energia projects",
   "license": "AGPL-3.0",
   "homepage": "https://som-energia.github.io/somenergia-ui/",
@@ -94,5 +94,9 @@
     "vite": "^4.4",
     "vite-plugin-svgr": "^4.2.0",
     "vitest": "^1.6.0"
+  },
+  "engines": {
+    "node": ">=18.0.0 <25.0.0",
+    "npm": ">=8.0.0"
   }
 }
